Add a Web Ext native messaging bridge.
The bridge communicates with the Web Extension via length-prefixed JSON
messages over std{in, out}. It manages a single `Toodle` store, in a
fixed location, and translates the messages into method calls. The API
still needs some work: in particular, we always read before and after
writing; we can't set labels yet, because we need to fetch their IDs to
set them on items, and removing items and labels isn't supported yet.
Error handling and reporting could also be more robust.

For the bridge to work, the native manifest must live in a specific
location, and include a `path` key that points to the bridge binary.
Running `./install-native-manifest.js` automatically sets the path and
copies the manifest to the correct location on macOS. Windows and Linux
use different locations, but the install script doesn't handle them
yet. This only needs to be done once.