Attempt to get toodle working with a recent mentat.
This isn't actually a PR, because it doesn't actually work :)
However, putting it up as a PR seems a good way to get feedback.
It also isn't strictly necessary to keep toodle alive in this way - as
Emily mentioned, simply pinning it to an earlier mentat revision would
work, but I took this on as an opportunity to better understand mentat.

The biggest set of changes here seems to be due to mentat wrapping a
TypedValue in a Binding enum. The simplest fix to that is to use the
.val() method on a binding, but that returns an Option, but many things
that want to use the TypedValue don't handle options - so I simply use
.val().expect() and hope nothing panics :) Is there a better way to handle
this?

The next issue I struck was due to toodle.initialize() not being called in
any of the tests. I'm a bit confused by that because it's not clear to me
how mentat changes would affect that.

And finally, while most tests now pass, all tests which end up calling
fetch_labels_for_item() fail - it always returns zero items. Best I can tell,
the into_rel_result() call there is returning a RelResult with a width of 3
but an empty vec - so it's not working even before my dodgy
/rows.iter()/rows.into_iter()/ change in that same function.

I don't mind poking more at this but also don't mind abandoning it either,
but I would appreciate a little feedback on the changes I made to help me
better understand what is going on.